bug fix for survey package 4.1
an update to svyquantile broke existing code. The new svyquantile
object is different from the older versions of it and it also
does not have an as.data.frame message. So I had to get elements
from the new svyquantile object, move them around, and convert them
to a data frame.